IOSS: Consolidate exodus database classes into single directory
