+++ packages/seacas/libraries/ioss/src/exo_fac/CMakeLists.txt
@@ -11,10 +11,7 @@ INCLUDE_DIRECTORIES(
   "${CMAKE_CURRENT_BINARY_DIR}/../"
 )
 
-SET(IOSS_DEPENDENCIES Iofx)
-IF (TPL_Netcdf_PARALLEL)
-  APPEND_SET(IOSS_DEPENDENCIES Iopx)
-ENDIF()
+SET(IOSS_DEPENDENCIES Ioex)
 
 TRIBITS_ADD_LIBRARY(
 	Ioexo_fac
