--- packages/seacas/libraries/ioss/src/exo_fpp/CMakeLists.txt
@@ -1,23 +1,0 @@
-
-SET(HEADERS "")
-SET(SOURCES "")
-
-SET_AND_INC_DIRS(DIR ${CMAKE_CURRENT_SOURCE_DIR})
-APPEND_GLOB(HEADERS ${DIR}/*.h)
-APPEND_GLOB(SOURCES ${DIR}/*.C)
-
-INCLUDE_DIRECTORIES(
-  "${CMAKE_CURRENT_SOURCE_DIR}/../"
-  "${CMAKE_CURRENT_BINARY_DIR}/../"
-)
-
-TRIBITS_ADD_LIBRARY(
-	Iofx
-	HEADERS	${HEADERS}
-	SOURCES ${SOURCES}
-	DEPLIBS Ioss Ioex
-)
-
-IF (BUILD_TESTING)
-  ENABLE_TESTING()
-ENDIF (BUILD_TESTING)
