--- packages/seacas/libraries/ioss/src/exo_fpp/cmake/Dependencies.cmake
@@ -1,5 +1,0 @@
-TRIBITS_PACKAGE_DEFINE_DEPENDENCIES(
-  LIB_REQUIRED_PACKAGES SEACASExodus
-  LIB_OPTIONAL_TPLS HDF5
-)
-
