--- packages/seacas/libraries/ioss/src/exo_par/CMakeLists.txt
@@ -1,34 +1,0 @@
-
-SET(HEADERS "")
-SET(SOURCES "")
-
-ASSERT_DEFINED(${PROJECT_NAME}_ENABLE_Zoltan)
-IF (NOT ${PROJECT_NAME}_ENABLE_Zoltan)
-  ADD_DEFINITIONS(-DNO_ZOLTAN_SUPPORT)
-ENDIF()
-
-IF (${PROJECT_NAME}_ENABLE_ParMETIS)
-ELSE()
-  ADD_DEFINITIONS(-DNO_PARMETIS_SUPPORT)
-ENDIF()
-
-SET_AND_INC_DIRS(DIR ${CMAKE_CURRENT_SOURCE_DIR})
-APPEND_GLOB(HEADERS ${DIR}/*.h)
-APPEND_GLOB(SOURCES ${DIR}/*.C)
-
-INCLUDE_DIRECTORIES(
-  "${CMAKE_CURRENT_SOURCE_DIR}/../"
-  "${CMAKE_CURRENT_BINARY_DIR}/../"
-  "${Zoltan_INCLUDE_DIRS}"
-  )
-
-TRIBITS_ADD_LIBRARY(
-  Iopx
-  HEADERS ${HEADERS}
-  SOURCES ${SOURCES}
-  DEPLIBS Ioss Ioex
-  )
-
-IF (BUILD_TESTING)
-  ENABLE_TESTING()
-ENDIF (BUILD_TESTING)
