+++ packages/seacas/libraries/ioss/src/exodus/CMakeLists.txt
@@ -2,9 +2,25 @@
 SET(HEADERS "")
 SET(SOURCES "")
 
+ASSERT_DEFINED(${PROJECT_NAME}_ENABLE_Zoltan)
+IF (NOT ${PROJECT_NAME}_ENABLE_Zoltan)
+  ADD_DEFINITIONS(-DNO_ZOLTAN_SUPPORT)
+ENDIF()
+
+IF (${PROJECT_NAME}_ENABLE_ParMETIS)
+ELSE()
+  ADD_DEFINITIONS(-DNO_PARMETIS_SUPPORT)
+ENDIF()
+
 SET_AND_INC_DIRS(DIR ${CMAKE_CURRENT_SOURCE_DIR})
 APPEND_GLOB(HEADERS ${DIR}/*.h)
 APPEND_GLOB(SOURCES ${DIR}/*.C)
+
+IF (NOT TPL_ENABLE_MPI)
+  LIST(REMOVE_ITEM SOURCES ${DIR}/Ioex_ParallelDatabaseIO.C ${DIR}/Ioex_DecompositionData.C)
+  LIST(REMOVE_ITEM HEADERS ${DIR}/Ioex_ParallelDatabaseIO.h ${DIR}/Ioex_DecompositionData.h)
+ENDIF()
+
 
 INCLUDE_DIRECTORIES(
   "${CMAKE_CURRENT_SOURCE_DIR}/../"
