+++ packages/seacas/libraries/ioss/src/init/CMakeLists.txt
@@ -22,10 +22,6 @@ ENDIF()
 IF (${PACKAGE_NAME}_ENABLE_SEACASExodus)
   SET(EXODUSLIB Ioex)
   SET(EXODUSFAC Ioexo_fac)
-
-  IF (TPL_Netcdf_PARALLEL)
-    SET(PAREXOLIB Iopx)
-  ENDIF()
 ENDIF()
 
 IF (TPL_ENABLE_CGNS)
@@ -44,7 +40,7 @@ TRIBITS_ADD_LIBRARY(
 	Ionit
 	HEADERS	${HEADERS}
 	SOURCES ${SOURCES}
-	DEPLIBS Ioss ${EXODUSLIB} ${PAREXOLIB}
+	DEPLIBS Ioss ${EXODUSLIB}
 	        Iogn Iogs Iohb Iotr ${IOVISLIB} ${EXODUSFAC}
             ${PAMGENLIB} ${CGNSLIB} ${VISLIB} ${DATAWAREHOUSELIB} ${ADIOS2LIB}
 )
