+++ packages/seacas/libraries/ioss/src/unit_tests/UnitTestIofxDatabaseIO.C
@@ -1,4 +1,4 @@
-// Copyright(C) 1999-2017 National Technology & Engineering Solutions
+// Copyright(C) 1999-2017, 2020 National Technology & Engineering Solutions
 // of Sandia, LLC (NTESS).  Under the terms of Contract DE-NA0003525 with
 // NTESS, the U.S. Government retains certain rights in this software.
 //
@@ -36,7 +36,7 @@
 #include <Ioss_PropertyManager.h>
 #include <Ioss_Region.h>
 #include <Ioss_Shell4.h>
-#include <exo_fpp/Iofx_DatabaseIO.h>
+#include <exodus/Ioex_DatabaseIO.h>
 #include <gtest/gtest.h>
 #include <mpi.h>
 #include <string>
@@ -59,7 +59,7 @@ namespace {
                      const int parameteric_dim, const int num_vertices, const int num_nodes,
                      const int num_edges, const int num_faces, const int num_boundaries);
 
-  Iofx::DatabaseIO *create_input_db_io(const std::string &filename)
+  Ioex::DatabaseIO *create_input_db_io(const std::string &filename)
   {
     Ioss::Init::Initializer init_db;
 
@@ -70,13 +70,13 @@ namespace {
     properties.add(Ioss::Property("INTEGER_SIZE_DB", 8));
     properties.add(Ioss::Property("INTEGER_SIZE_API", 8));
 
-    Iofx::DatabaseIO *db_io =
-        new Iofx::DatabaseIO(nullptr, filename, db_usage, communicator, properties);
+    Ioex::DatabaseIO *db_io =
+        new Ioex::DatabaseIO(nullptr, filename, db_usage, communicator, properties);
     return db_io;
   }
 
   // BeginDocTest1
-  Iofx::DatabaseIO *create_output_db_io(const std::string &filename)
+  Ioex::DatabaseIO *create_output_db_io(const std::string &filename)
   {
     Ioss::Init::Initializer init_db;
 
@@ -87,15 +87,15 @@ namespace {
     properties.add(Ioss::Property("INTEGER_SIZE_DB", 8));
     properties.add(Ioss::Property("INTEGER_SIZE_API", 8));
 
-    Iofx::DatabaseIO *db_io =
-        new Iofx::DatabaseIO(nullptr, filename, db_usage, communicator, properties);
+    Ioex::DatabaseIO *db_io =
+        new Ioex::DatabaseIO(nullptr, filename, db_usage, communicator, properties);
     return db_io;
   }
 
-  TEST(Iofx, test_constructor)
+  TEST(Ioex, test_constructor)
   {
     const std::string filename = "ADeDA.e";
-    Iofx::DatabaseIO *db_io    = create_input_db_io(filename);
+    Ioex::DatabaseIO *db_io    = create_input_db_io(filename);
 
     Ioss::Region region(db_io);
 
@@ -277,10 +277,10 @@ namespace {
   }
 
   // BeginDocTest2
-  TEST(Iofx, test_writing_of_file)
+  TEST(Ioex, test_writing_of_file)
   {
     const std::string input_filename = "ADeDA.e";
-    Iofx::DatabaseIO *db_in          = create_input_db_io(input_filename);
+    Ioex::DatabaseIO *db_in          = create_input_db_io(input_filename);
 
     Ioss::Region input_region(db_in);
 
@@ -288,7 +288,7 @@ namespace {
 
     const std::string output_filename = "ADeDA_out.e";
 
-    Iofx::DatabaseIO *db_out = create_output_db_io(output_filename);
+    Ioex::DatabaseIO *db_out = create_output_db_io(output_filename);
     Ioss::Region      output_region(db_out);
     EXPECT_TRUE(db_out->ok());
